H-OBS-R-SPAR/requirements.txt
# Core Deep Learning Framework
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0

# CUDA and GPU utilities
nvidia-ml-py3==7.352.0
pycuda==2022.2.2

# Reinforcement Learning
stable-baselines3==2.2.1
gym==0.26.2
tensorboard==2.15.1

# Hessian Approximation & Optimization
backpack-for-pytorch==1.6.0
pytorch-ignite==0.4.13

# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.7.5
seaborn==0.13.0
plotly==5.18.0
tensorboardX==2.6.2.2

# Statistical Analysis
statsmodels==0.14.1
pingouin==0.5.3

# Dataset & Model Zoo
timm==0.9.12
torchmetrics==1.3.0

# System Profiling
gpustat==1.1.1
psutil==5.9.7
py3nvml==0.2.7

# TensorRT Integration
tensorrt==8.6.1
onnx==1.15.0
onnx-simplifier==0.4.35

# Configuration Management
pyyaml==6.0.1
hydra-core==1.3.2
omegaconf==2.3.0

# Utilities
tqdm==4.66.1
colorama==0.4.6
rich==13.7.0
loguru==0.7.2

# Testing & Reproducibility
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0

# Pruning Library
torch-pruning>=1.3.0
